Rename SessionsDataView.h/.cpp to PresetsDataView.h/.cpp
