+++ OrbitGl/App.cpp
@@ -38,13 +38,13 @@
 #include "OrbitSession.h"
 #include "Params.h"
 #include "Pdb.h"
+#include "PresetsDataView.h"
 #include "PrintVar.h"
 #include "ProcessesDataView.h"
 #include "SamplingProfiler.h"
 #include "SamplingReport.h"
 #include "SamplingUtils.h"
 #include "ScopeTimer.h"
-#include "SessionsDataView.h"
 #include "StringManager.h"
 #include "TextRenderer.h"
 #include "Utils.h"
