+++ OrbitGl/App.h
@@ -29,10 +29,10 @@
 #include "OrbitBase/MainThreadExecutor.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/ThreadPool.h"
+#include "PresetsDataView.h"
 #include "ProcessManager.h"
 #include "ProcessesDataView.h"
 #include "SamplingReportDataView.h"
-#include "SessionsDataView.h"
 #include "StringManager.h"
 #include "SymbolHelper.h"
 #include "Threading.h"
